Fix chan_ooh323 so it works on Mac OS X, as well.
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@254799 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ addons/ooh323c/src/ootrace.c
@@ -111,7 +111,7 @@ void ooTraceLogMessage(const char * logMessage)
       struct timeval systemTime;
       gettimeofday(&systemTime, NULL);
       fprintf(gH323ep.fptraceFile, "%s:%03ld  %s", timeString, 
-               systemTime.tv_usec/1000, logMessage);
+               (long) systemTime.tv_usec/1000, logMessage);
    }
    else
       fprintf(gH323ep.fptraceFile, "%s", logMessage);
